fetch package refactor
