+++ src/main/java/com/configcat/ConfigCatClient.java
@@ -3,6 +3,7 @@ package com.configcat;
 import com.configcat.cache.ConfigCache;
 import com.configcat.cache.NullConfigCache;
 import com.configcat.evaluation.*;
+import com.configcat.fetch.ConfigFetcher;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogMessages;
 import com.configcat.log.ConfigCatLogger;
