+++ src/main/java/com/configcat/ConfigService.java
@@ -1,6 +1,8 @@
 package com.configcat;
 
 import com.configcat.cache.ConfigCache;
+import com.configcat.fetch.ConfigFetcher;
+import com.configcat.fetch.FetchResponse;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogMessages;
 import com.configcat.log.ConfigCatLogger;
