+++ src/main/java/com/configcat/Utils.java
@@ -3,8 +3,8 @@ package com.configcat;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
-final class Utils {
+public final class Utils {
     private Utils() { /* prevent from instantiation*/ }
 
-    static final Gson gson = new GsonBuilder().create();
+    public static final Gson gson = new GsonBuilder().create();
 }
