+++ src/test/java/com/configcat/AutoPollingPolicyTest.java
@@ -3,6 +3,7 @@ package com.configcat;
 import com.configcat.cache.ConfigCache;
 import com.configcat.cache.NullConfigCache;
 import com.configcat.cache.SingleValueCache;
+import com.configcat.fetch.ConfigFetcher;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
